Integrate test mode into imager
When the user selects the test mode in the UI, it inserts a standard
test_sources.json file into the sysconfig directory. This configuration
has one channel that points to the uploaded file.

The imager is modified to first look for test_sources.json before
sources.json. If found, it would load the former in lieu of sources.json.

The next time when  the user selects the streaming mode in the UI, it
deletes the test_sources.json file from the sysconfig directory to revert
the system to its standard streaming behavior.