orchestrator: vllm-complex interface, dataset tweaks
The vllm-complex interface is for using with vLLM serving a
fine-tuned model. The interface itself is open-ai compatible,
but the prompt is custom. It cannot be used with 3rd party
open-ai compatible servers serving LLAMA 3.2 Vision Instruct
without the custom chat template.

The dataset collection code is changed to capture negative
sample before and after detecting positive(s). This should
give it a chance to capture some false negatives.