VLLMComplexInterface: gracefully handle absence of the server
The VLLMComplexInterface tries to query the list of models
served by the service and pick one automatically if no model
name is configured manually. This was causing exceptions and
orchestratior failing to start if vLLM service was not running.
This change ignores the failures and attempts to retry the
automatic model name selection at the time of inference, to
allow strating vLLM after the watcman services are launched.