ui: manual dataset management
This commit adds UI for the dataset management, manual labeling
and aggregation of the collected data for the model fine tuning.

The positive and negative samples selected by the Watchman dataset
service are collected for each channel/object pair in the bucket
topped at 1000 samples. The bucket can be added to the labeling queue
where up to 9 such buckets can wait to be labeled.

The UI allows to label samples positive, negative or mark to be
skipped. No object location description tuning is supported at the
moment (i.e. the training is done only for the simple object
present detection prompt).

The data from the labeled buckets can then be aggregated in the .pkl
data file and used for the model fine tuning in the fine_tune.ipynb
notebook.