ui: added ability to auto-label the dataset images
The dataset images can now be auto-labeled using any of the watchman's
model interfaces. The openai-generic interface was added to allow using
the OpenAI's online service and automate the manual labeling.
The cost to go through a 1000 of 1280x720 images and label with only the
object presence detection (no verbal description of the object location)
using o4-mini model was $2.50.